LSTM model
The predictive model which I used is LSTM. Although it does not give as much good results as it should be but I used this because there were some restriction on using libraries, that's why I came up with this one. The model predicts spread of virus in Pakistan using those countries data whose trend of spread was almost similar to the trend of spread of Pakistan.